refactor: task-6
