+++ shared/package.json
